+++ Utilities/DebugConsole/DebugConsole.cs
@@ -1,6 +1,7 @@
 ﻿using Base.Core;
 using Base.Utils.GameConsole;
 using Harmony;
+using Newtonsoft.Json;
 using PhoenixPoint.Home.View.ViewModules;
 using PhoenixPoint.Home.View.ViewStates;
 using System;
@@ -16,13 +17,14 @@ using UnityEngine;
 namespace Sheepy.PhoenixPt.DebugConsole {
 
    public class ModConfig {
-      public int  Config_Version = 20200324;
+      public int  Config_Version = 20200405;
       public bool Mod_Count_In_Version = true;
       public bool Log_Game_Error = true;
       public bool Log_Game_Info = true;
       public bool Log_Modnix_Error = true;
       public bool Log_Modnix_Info = true;
       public bool Log_Modnix_Verbose = false;
+      public bool Scan_Mods_For_Command = true;
       public bool Write_Modnix_To_Console_Logfile = false;
    }
 
@@ -34,10 +36,16 @@ namespace Sheepy.PhoenixPt.DebugConsole {
       public void SplashMod ( Func< string, object, object > api = null ) {
          GameConsoleWindow.DisableConsoleAccess = false; // Enable console first no matter what happens
          SetApi( api, out Config );
+         CheckConfig();
          Verbo( "Console Enabled" );
          if ( Config.Log_Game_Error || Config.Log_Game_Info ) {
-            Application.logMessageReceived += UnityToConsole;
+            Application.logMessageReceived += UnityToConsole; // Non-main thread will cause inconsistent formats.
             TryPatch( typeof( TimingScheduler ), "Update", postfix: nameof( BufferToConsole ) );
+         }
+         if ( Config.Scan_Mods_For_Command ) {
+            TryPatch( typeof( ConsoleCommandAttribute ), "GetCommands", prefix: nameof( ScanCommands ) );
+            TryPatch( typeof( ConsoleCommandAttribute ), "GetInfo", prefix: nameof( ScanCommands ) );
+            TryPatch( typeof( ConsoleCommandAttribute ), "HasCommand", prefix: nameof( ScanCommands ) );
          }
          if ( api == null ) return;
 
@@ -58,6 +66,69 @@ namespace Sheepy.PhoenixPt.DebugConsole {
                Info( "Modnix assembly not found, log not forwarded." );
          }
       }
+
+      private static void CheckConfig () {
+         if ( Config.Config_Version < 20200405 )
+            ModnixApi?.Invoke( "config_save", Config );
+      }
+
+      private static HashSet< Assembly > ScannedMods;
+      private static FieldInfo CmdMethod;
+      private static FieldInfo CmdVarArg;
+
+      private static void InitScanner () {
+         if ( ScannedMods != null ) return;
+         ScannedMods = new HashSet< Assembly >();
+         CmdMethod = typeof( ConsoleCommandAttribute ).GetField( "_methodInfo", BindingFlags.NonPublic | BindingFlags.Instance );
+         CmdVarArg = typeof( ConsoleCommandAttribute ).GetField( "_variableArguments", BindingFlags.NonPublic | BindingFlags.Instance );
+      }
+
+      private static void ScanCommands ( SortedList<string, ConsoleCommandAttribute> ___CommandToInfo ) { try {
+         InitScanner();
+         Assembly[] asmAry = AppDomain.CurrentDomain.GetAssemblies();
+         if ( asmAry.Length == ScannedMods.Count ) return;
+         foreach ( var asm in asmAry ) {
+            if ( ScannedMods.Contains( asm ) ) continue;
+            if ( asm.FullName.StartsWith( "UnityEngine.", StringComparison.Ordinal ) ||
+                 asm.FullName.StartsWith( "Unity.", StringComparison.Ordinal ) ||
+                 asm.FullName.StartsWith( "Assembly-CSharp,", StringComparison.Ordinal ) ||
+                 asm.FullName.StartsWith( "System.", StringComparison.Ordinal ) ) {
+               ScannedMods.Add( asm );
+               continue;
+            }
+            Verbo( "Scanning {0} for console commands.", asm.FullName );
+            foreach ( var type in asm.GetTypes() )
+               foreach ( var func in type.GetMethods( BindingFlags.Static | BindingFlags.Public ) ) {
+                  var tag = func.GetCustomAttribute( typeof(ConsoleCommandAttribute) ) as ConsoleCommandAttribute;
+                  if ( tag == null ) continue;
+                  tag.Command = tag.Command ?? func.Name;
+                  if ( ___CommandToInfo.ContainsKey( tag.Command ) ) {
+                     Info( "Command exists, cannot register {0} of {1} in {2}.", tag.Command, type.FullName, asm.FullName );
+                     continue;
+                  }
+                  CmdMethod.SetValue( tag, func );
+                  var param = func.GetParameters();
+                  if ( param.Length > 0 && param[ param.Length - 1 ].ParameterType.FullName.Equals( "System.String[]" ) )
+                     CmdVarArg.SetValue( tag, true );
+                  ___CommandToInfo.Add( tag.Command, tag );
+                  Info( "Command registered: {0} of {1} in {2}.", tag.Command, type.FullName, asm.FullName );
+               }
+            ScannedMods.Add( asm );
+         }
+      } catch ( Exception ex ) { Error( ex ); } }
+
+      [ ConsoleCommand( Command = "modnix", Description = "Call Modnix or compatible api." ) ]
+      public static void ApiCommand ( IConsole console, string[] param ) { try {
+         if ( param == null || param.Length == 0 ) throw new ApplicationException( "Api action required." );
+         object arg = null;
+         if ( param.Length > 2 ) {
+            arg = new string[ param.Length - 1 ];
+            Array.Copy( param, 1, arg as string[], 0, param.Length - 1 );
+         } else if ( param.Length == 1 )
+            arg = param[1];
+         var result = ModnixApi( param[0], arg );
+         console.WriteLine( EscLine( result is string txt ? txt : JsonConvert.SerializeObject( result, Formatting.None ) ) );
+      } catch ( Exception ex ) { Error( ex ); } }
 
       private static void AfterMainMenu_AddModCount ( UIStateMainMenu __instance ) { try {
          var revision = typeof( UIStateMainMenu ).GetProperty( "_buildRevisionModule", BindingFlags.NonPublic | BindingFlags.Instance )?.GetValue( __instance ) as UIModuleBuildRevision;
@@ -81,22 +152,22 @@ namespace Sheepy.PhoenixPt.DebugConsole {
             lock ( _Lock ) switch ( level ) {
                case TraceEventType.Critical: case TraceEventType.Error:
                   if ( ! Config.Log_Modnix_Error ) return;
-                  prefix = "Error";
+                  prefix = "<color=red>Error";
                   break;
                 case TraceEventType.Warning:
                   if ( ! Config.Log_Modnix_Error ) return;
-                  prefix = "Warning";
+                  prefix = "<color=orange>Warning";
                   break;
                case TraceEventType.Information :
                   if ( ! Config.Log_Modnix_Info ) return;
-                  prefix = "Log [INFO]";
+                  prefix = "<color=lime>Log [INFO]";
                   break;
                default :
                   if ( ! Config.Log_Modnix_Verbose ) return;
-                  prefix = "Log [VEBO]";
+                  prefix = "<color=aqua>Log [VEBO]";
                   break;
             }
-            txt = $"{prefix} {Time.frameCount} ({Time.time.ToString("0.000")}) {txt}";
+            txt = $"{prefix} {Time.frameCount} ({Time.time.ToString("0.000")}) {txt}</color>";
          }
          lock ( Buffer ) Buffer.Add( txt );
       } catch ( Exception ex ) { Error( ex ); } }
@@ -113,10 +184,11 @@ namespace Sheepy.PhoenixPt.DebugConsole {
             default:
                if ( ! Config.Log_Game_Info ) return; break;
          }
-         // Long message can cause the 
+         // Skip long message which can cause the 65000 vertices error.
          if ( condition.Length > 1000 || stackTrace.Length > 1000 ||
               condition.Contains( "Called from a secondary Thread" ) ||
-              stackTrace?.Contains( "UnityTools.LogFormatter" ) == true ) return;
+              stackTrace?.Contains( "UnityTools.LogFormatter" ) == true )
+           return;
          var line = $"{type} {condition}   {stackTrace}".Trim();
          if ( line.Length == 0 ) return;
          lock ( Buffer ) Buffer.Add( line );
@@ -131,7 +203,9 @@ namespace Sheepy.PhoenixPt.DebugConsole {
          }
          var console = GameConsoleWindow.Create();
          foreach ( var line in lines )
-            console.WriteLine( line );
+            console.WriteLine( EscLine( line ) );
       } catch ( Exception ex ) { Error( ex ); } }
+
+      private static string EscLine ( string line ) => line.Replace( "{", "{{" ).Replace( "}", "}}" );
    }
 }