+++ Utilities/DebugConsole/mod_info.js
@@ -13,6 +13,9 @@
       "Nexus"  : "https://nexusmods.com/phoenixpoint/mods/44/",
       "GitHub" : "https://github.com/Sheep-y/PhoenixPt-Mods"
    },
+   ConfigType: "Sheepy.PhoenixPt.DebugConsole.ModConfig",
+
+   /* Legacy config for Modnix 1.0 */
    DefaultConfig: {
       "Config_Version" : 20200324,
       "Mod_Count_In_Version" : true,
@@ -21,6 +24,7 @@
       "Log_Modnix_Error" : true,
       "Log_Modnix_Info" : true,
       "Log_Modnix_Verbose" : false,
+      "Scan_Mods_For_Command" : true,
       "Write_Modnix_To_Console_Logfile" : false,
    },
 })